requirements.txt
# AgenticSeek Python Dependencies
# Core dependencies
httpx>=0.25.0
anthropic>=0.20.0
openai>=1.5.0
python-dotenv>=1.0.0
redis>=5.0.0
pydantic>=2.5.0

# Web scraping and automation
selenium>=4.15.0
undetected-chromedriver>=3.5.4
beautifulsoup4>=4.12.0
requests>=2.31.0

# Async and concurrency
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# Audio processing (optional)
kokoro>=0.1.0  # Text-to-speech
speechrecognition>=3.10.0  # Speech-to-text

# Utilities
pyyaml>=6.0.1
click>=8.1.7
tqdm>=4.66.0
colorama>=0.4.6

# Monitoring and logging
structlog>=23.2.0
rich>=13.7.0

# Security
cryptography>=41.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0